Compile VS solution w/o warnings other than C4251
Warning C4251 complains about template classes needing
dll-interfaces. This issue needs a separate fix.
Some warning fixes will merely apply to building with the
solution shipped with virvo (e.g. MS deprecation warnings
for C runtime functions and some posix functions)

git-svn-id: https://deskvox.svn.sourceforge.net/svnroot/deskvox/trunk@1078 04231f92-3938-0410-9931-e931ac552b4f

+++ virvo/virvo/vvbsptree.cpp
@@ -115,7 +115,7 @@ void vvHalfSpace::setBrickList(const std::vector<BrickList>& brickList)
   vvVector3 minCorner = vvVector3(VV_FLT_MAX, VV_FLT_MAX, VV_FLT_MAX);
   vvVector3 maxCorner = vvVector3(-VV_FLT_MAX, -VV_FLT_MAX, -VV_FLT_MAX);
 
-  for (int f=0; f<brickList.size(); ++f)
+  for (size_t f=0; f<brickList.size(); ++f)
   {
     for (BrickList::const_iterator it = brickList[f].begin(); it != brickList[f].end(); ++it)
     {
@@ -224,7 +224,7 @@ float vvHalfSpace::calcContainedVolume() const
   float w, h, d;
 
   w = h = d = 0.0f;
-  for (int f=0; f<_brickList.size(); ++f)
+  for (size_t f=0; f<_brickList.size(); ++f)
   {
     for(BrickList::const_iterator it = _brickList[f].begin(); it != _brickList[f].end(); ++it)
     {
@@ -533,7 +533,7 @@ vvHalfSpace* vvSpacePartitioner::getAABBHalfSpaces(const std::vector<BrickList>&
   result[0].setBrickList(std::vector<BrickList>());
   result[1].setBrickList(std::vector<BrickList>());
 
-  for (int f=0; f<brickList.size(); ++f)
+  for (size_t f=0; f<brickList.size(); ++f)
   {
     result[0].getBrickList().push_back(BrickList());
     result[1].getBrickList().push_back(BrickList());
